requirements.txt
streamlit>=1.28.0
matplotlib>=3.7.0
plotly>=5.15.0
google-generativeai>=0.3.0
lxml>=4.9.0
Pillow>=10.0.0
pandas>=2.0.0